Generate custom `winmd` from headers with improved annotations
The winmd shipped with the NuGet package has an [indecisive license][1],
as it's intended to fall under the MIT license just like other code
files, but is strictly a binary blob.  At the same time the metadata
[isn't representing the headers as conveniently and completely][2] as
we'd like in our Rust bindings, but unfortunately there is no upstream
source available to reference or contribute to, where we can tweak and
improve how the NuGet `winmd` is generated.

The metadata description introduced here tries to reproduce the original
metadata as closely as possible, while following the improvements
[that I requested upstream][2] and [maintaining the `int` type for
`DSTORAGE_DEBUG`][3].

[1]: https://github.com/microsoft/DirectStorage/issues/8#issuecomment-1875760330
[2]: https://github.com/microsoft/DirectStorage/issues/38
[3]: https://github.com/microsoft/win32metadata/issues/1822#issuecomment-1894407892