metadata: One-line `TraverseFiles` to parse `#define`s from `dstorageerr.h`
Some manual fiddling with `generate.proj` showed to me that only the
first file in `TraverseFiles` was getting considered for constants
scraping. There doesn't appear to be a `.Trim()` call on the string
parsing in `WinmdGenerator` source code, likely causing the last entry
to end with `\n` which won't match a file name. Simply changing how
`TraverseFiles` is specified in the project file XML allows us to remove
the manual definitions of these constants entirely again.