Allow using `windows-targets` `0.52` next to `0.53`
In #26 we bumped `windows-bindgen` to `0.59` which now generates
`#[link] extern "system" { fn ... }` blocks via the `link!()` macro in
the `windows_targets` crate.  I quickly added a dependency on `0.53`
there, but the macro is also available in `0.52` which is also still
used by `windows 0.57-0.58`.  By not allowing `direct-storage` to use
`windows-bindgen 0.52`, the user always has to accept a crate duplicate
until they are ready to upgrade to `windows 0.59`.  Turn this into a
`>=0.52, <=0.53` version range to support either revision of the crate.