Bump `windows(-core)` crate range to `0.57-0.60`
https://github.com/microsoft/windows-rs/releases/tag/0.62.0

This bumps the MSRV to 1.71 for using the new `windows-link` crate
over `windows-targets`, and I'm not sure how that fares when the
older `windows(-core)` crates allowed by our version range still use
`windows-targets`.

We could set `--link windows_targets` to opt out of this change for now,
or perhaps have a feature to opt everything in to `windows-link`?