Upgrade `windows-bindgen` to `0.64` and "drop" `windows-link` dependency
http://github.com/Microsoft/windows-rs/releases/70

Following a discussion on `windows-bindgen` and subsequent change
in https://togithub.com/microsoft/windows-rs/pull/3763 the resulting
bindings now try to access as many (reexported) bindings directly
via `windows_core` rather than requiring all transitive deps
like `windows-link` (and in other crates: `windows-collections`,
`windows-result`, `windows-strings`...) to be explicitly specified in
`Cargo.toml`.  This is even more confusing and hard to manage because of
version range dependencies, where the ranges have to line up across all
of them.

Unfortunately `windows-link::link!()` wasn't reexported from
`windows-core` until `0.62.1`, making us no longer compatible with
`0.61` if this is merged.

Note that a separate request is still pending to do the same
for the `windows` crate if it is needed next to `windows_core`,
so that the `windows::core::` reexport can be reused whenever
possible and obviate the need for a separate `windows-core
= "range"` dependency that must match `windows = "range"`:
https://togithub.com/microsoft/windows-rs/issues/3776